Successful migration to CreateJS
After considering multiple libraries for the game, CreateJS was chosen
as the best fit all the game needs. EasleJS will be used to draw the
game assets, TweenJs will be used for the animations, SoundJS can be
used to implement game audio and if needed PreloadJS can be used to
improve load times.