Restructure graphic attributes into obj and func
This is a major restructure of the graphical elements of the game.
Graphical attributes have been wrapped into an object as the
namespace was getting crowded. The object stores all unchanging
attributes for each game room such as the pillar widths, heights, and
locations. All the sizes of graphical objects are relative to the size
of Simon, the multiplier, and the screen size so it makes it easier to
follow the relationships between them since they are now all in one
place. Also in this commit a lot of the calculations for the layout of
each game room have been moved to functions so that the attributes only
need be calculated once per room such as when the next move is added.
Before, they were recalculated every time the screen is redrawn and
while they aren't expensive calculations I presume the effects would be
noticed when the animation is implemented. I also moved the gameState
object to simon-logic as that object tracks the game logic values.