Add basic animation loop with move queuing
A main loop function has been started for all animation in the game. The
loop is executed every CreateJS tick and depending on the game state
different animations are played. The game now queues up moves to be
animated so that one animation can finish before the next one starts.
This is done to ensure that the user can enter in moves at any point
during an animation without triggering the next one which before lead to
skipped animations. The loop is in a basic but semifunctional state as
of now but significant work needs to be done. Serious bugs need to be
removed and many more animations need to be added but this is a
functional start.