Smooth animation with screen scrolling
In this commit the animations are improved and screen scrolling is
implemented. Once simon gets to the center of the screen the room around
him scrolls so that content outside the canvas view can scroll into
view. The jump animations now follow a curve rather than a straight line
making for a more pleasing animation. A temporary solution for
dynamically resizing the screen mid game is also implemented. Previously
if the window was resized mid game it could not handle redrawing the
current game so for now a gameover is thrown and the game resets.