Add animations for all hazards, rename functions
Animations for all four hazards are added, Up, Down, Left, and Right.
The checkGame function is renamed gameCheck to better align with other
function names and the 'extra user move' checking part of the code has
been moved out to it's own function, gameOverflow. In doing this it now
allows extra moves that are added by the user to be removed during the
game tick before they can cause any issues later.