Add hazard objects with animations
The hazard objects are added to the animation at every move. The
functions to draw the hazards are also added, where up uses a bolt of
electricity, down uses a spike plate, and left and right use a single
spike. Each hazard is dynamically generated when a move is made and
stored in a container in the room container. Some comment and general
tidying up has also been done.