Add title/gameover screen, invoke responsiveness
A game over screen and title screen were added. In animating the game
over screen, should the user enter the wrong move (ie up instead of
down), simon does not move and the correct hazard animates before the
'stage' cuts to black. For the title screen, to use a larger font for
the title the textbox function was modified to take in a fontsize
argument allowing for variable text size.

The game should now be responsive using the multiplier flag in the
grphState object. The game should now scale appropriately with
screensize. The game also now checks if the orientation of the screen is
portrait rather than landscape (ie taller than it is wide) and if so
asks the user to rotate the screen. As per usual some comment and
general tidying has also been done