Make changes to remove linter errors
Started using JSHint, html-tidy and csslint directly in sublime to
better keep track of errors as they happen. To address warnings JSHint
gives about variables not defined in the scope of the js file I added
lines declaring the variables as global. Also reordered the css so that
the rules are in alphabetical order, didn't know that mattered!.