Add final fail state and solve mystery
Added the final fail state where if the user enters any other move than
right when jumping from pillar to pillar the game now fails with an
animation and custom dialogue. To use the custom dialogue the gameover
functions can now have custom text passed into it to animate at the end.
I also solved a longstanding mystery as to why the code would fail if
any lines were added to the gameCheck function. Turns out you can't have
lines between an 'if' and subsequent 'else if' statement! This took so
long to debug because the error would be thrown from another part of the
code so it wasn't immediately obvious what the issue was. Also there was
some general tiding up of some comments.