Add sound, break out graphics, imp. Prettier rule
Sounds have been added to the game using a combination of the SoundJS
library and the Audiosynth library. Now, when the game is started and on
each move, the notes from the physical Simon game play. These notes,
rather than needing a dedicated sound file, are generated on the fly
using the Audiosynth library. Using this library means that there can be
better control over the duration of the notes and should hopefully mean
the game loads faster not needing extra files. To add to the experience,
sound effects are also used for each of the hazards which have been
tuned to match their associated note. Those files are loaded using the
SoundJS library in the new file simon-audio.js.

To make the simon-graphics.js file more readable, the functions that are
used to draw the objects in the game have been broken out into their own
file, simon-draw.js. The event listener in simon-graphics.js that
handles redrawing the screen when the screen size changes has been
updated to fix a bug where the game over screen would not display.
Before, when the screen size changed, the function would fire multiple
times in quick succession meaning that the game over screen was
overwritten before it had a chance to load.

Tn simon-game.js the keyDownHandler has been simplified, removing the
need for the moveFlag variable and condition.

Finally in this update there has been some general cleaning up of the
code and comments and the code has been formatted using the Prettier
rule set. This has lead what seems like a lot of changes but the
majority of them are purely cosmetic.