Remove move queuing and reduce sound volume
After some users playtested the game it was found that the move queuing
system in the game was more confusing than helpful so for the moment it
has been turned off. The the system itself has not been removed, instead
while an animation plays the keyDownHandler will not accept new moves.
This is a temporary fix as to separate the system would require
significant work. Also in feedback some users found that the audio would
clip so the global audio has been reduced so that if any notes overlap
they are less likely to clip.