Complete README, add cheating and clean code
Adding the near finished version of the README, some sections may need
to be expanded on in the future but it should be pretty close. A feature
to help with testing the game was added where the game can now play
itself. Before, this was done by hand following the output on the
console but now if the cheaterFlag is set to true you can just sit back
and relax. Some areas of the code were also tweaked, removing deprecated
variables and cleaning up some of variable names.